on_select instaed of on_click
